Refactor functions into util files to test
